LCAM-950|Refactor schema in preparation for MAAT integration
