--- crown-court-contribution/src/main/resources/schemas/maat-api/apiCrownCourtSummary.json
@@ -1,16 +1,0 @@
-{
-  "$schema": "http://json-schema.org/draft-04/schema#",
-  "id": "apiCrownCourtSummary.json",
-  "type": "object",
-  "description": "Crown Court Summary",
-  "properties": {
-    "crownCourtOutcome": {
-      "type": "array",
-      "description": "list of crown court outcome",
-      "items": {
-        "$ref": "apiCrownCourtOutcome.json"
-      }
-    }
-  },
-  "additionalProperties": false
-}
